+++ Cargo.toml
@@ -1,48 +1,48 @@
 [package]
-authors = [ "KILT <info@kilt.io>" ]
-build = "build.rs"
-edition = "2018"
-name = "mashnet-node"
-version = "0.23.0"
+authors = [ 'KILT <info@kilt.io>' ]
+build = 'build.rs'
+edition = '2018'
+name = 'mashnet-node'
+version = '0.23.0'
 
 [[bin]]
-name = "mashnet-node"
-path = "src/main.rs"
+name = 'mashnet-node'
+path = 'src/main.rs'
 
 [profile.release]
-panic = "unwind"
+panic = 'unwind'
 
 [workspace]
-members = [ "runtime" ,
+members = [ 'runtime' , 'pallets/attestation', 'pallets/ctype', 'pallets/delegation', 'pallets/did', 'pallets/error', 'pallets/portablegabi'
 ]
 
 [build-dependencies]
-build-script-utils = { package = "substrate-build-script-utils", version = "2.0.0-rc5" }
-vergen = "3.1.0"
+build-script-utils = { package = 'substrate-build-script-utils', version = '2.0.0-rc5' }
+vergen = '3.1.0'
 
 [dependencies]
-frame-rpc-system = { package = "substrate-frame-rpc-system", version = "2.0.0-rc5" }
-futures = "0.3"
-sc-finality-grandpa = "0.8.0-rc5"
-sp-finality-grandpa = "2.0.0-rc5"
-hex-literal = "0.2"
-jsonrpc-core = "14.0.3"
-log = "0.4"
-mashnet-node-runtime = { path = "runtime" }
-sc-basic-authorship = "0.8.0-rc5"
-sc-cli = "0.8.0-rc5"
-sc-client-api = "2.0.0-rc5"
-sc-consensus = "0.8.0-rc5"
-sc-consensus-aura = "0.8.0-rc5"
-sc-executor = "0.8.0-rc5"
-sc-network = "0.8.0-rc5"
-sc-rpc = "2.0.0-rc5"
-sc-rpc-api = "0.8.0-rc5"
-sc-service = "0.8.0-rc5"
-sc-transaction-pool = "2.0.0-rc5"
-sp-consensus = "0.8.0-rc5"
-sp-consensus-aura = "0.8.0-rc5"
-sp-core = "2.0.0-rc5"
-sp-inherents = "2.0.0-rc5"
-sp-runtime = "2.0.0-rc5"
-structopt = "0.3.8"
+frame-rpc-system = { package = 'substrate-frame-rpc-system', version = '2.0.0-rc5' }
+futures = '0.3'
+sc-finality-grandpa = '0.8.0-rc5'
+sp-finality-grandpa = '2.0.0-rc5'
+hex-literal = '0.2'
+jsonrpc-core = '14.0.3'
+log = '0.4'
+mashnet-node-runtime = { path = 'runtime' }
+sc-basic-authorship = '0.8.0-rc5'
+sc-cli = '0.8.0-rc5'
+sc-client-api = '2.0.0-rc5'
+sc-consensus = '0.8.0-rc5'
+sc-consensus-aura = '0.8.0-rc5'
+sc-executor = '0.8.0-rc5'
+sc-network = '0.8.0-rc5'
+sc-rpc = '2.0.0-rc5'
+sc-rpc-api = '0.8.0-rc5'
+sc-service = '0.8.0-rc5'
+sc-transaction-pool = '2.0.0-rc5'
+sp-consensus = '0.8.0-rc5'
+sp-consensus-aura = '0.8.0-rc5'
+sp-core = '2.0.0-rc5'
+sp-inherents = '2.0.0-rc5'
+sp-runtime = '2.0.0-rc5'
+structopt = '0.3.8'
