+++ pallets/ctype/Cargo.toml
@@ -1,0 +1,66 @@
+[package]
+authors = [ 'KILT <info@kilt.io>' ]
+edition = '2018'
+name = 'ctype'
+repository = 'https://github.com/KILTprotocol/mashnet-node'
+version = '0.23.0'
+description = 'TODO'
+
+[package.metadata.docs.rs]
+targets = ['x86_64-unknown-linux-gnu']
+
+[build-dependencies.wasm-builder-runner]
+package = 'substrate-wasm-builder-runner'
+version = '1.0.6'
+
+[dependencies]
+error = { version = '0.23.0', default-features = false, path = '../error' }
+
+[dev-dependencies]
+mashnet-node-runtime = { version = '0.23.0', default-features = false, path = '../../runtime' }
+
+[dev-dependencies.sp-core]
+default-features = false
+version = '2.0.0-rc5'
+
+[dependencies.codec]
+default-features = false
+features = ['derive']
+package = 'parity-scale-codec'
+version = '1.3.0'
+
+[dependencies.support]
+default-features = false
+package = 'frame-support'
+version = '2.0.0-rc5'
+
+[dependencies.frame-system]
+default-features = false
+version = '2.0.0-rc5'
+
+[dependencies.sp-runtime]
+default-features = false
+version = '2.0.0-rc5'
+
+[dependencies.sp-std]
+default-features = false
+version = '2.0.0-rc5'
+
+[dependencies.runtime-io]
+default-features = false
+package = 'sp-io'
+version = '2.0.0-rc5'
+
+[features]
+default = [ 'std' ]
+std = [
+	'codec/std',
+	'support/std',
+	'frame-system/std',
+	'sp-runtime/std',
+	'runtime-io/std',
+	'sp-std/std',
+	'runtime-io/std',
+	'error/std'
+]
+
