+++ pallets/ctype/src/lib.rs
@@ -18,14 +18,15 @@
 
 //! CTYPE: Handles CTYPEs on chain,
 //! adding CTYPEs.
+#![cfg_attr(not(feature = "std"), no_std)]
 
 /// Test module for CTYPEs
 #[cfg(test)]
 mod tests;
 
-use super::error;
-use support::{debug, decl_event, decl_module, decl_storage, dispatch::DispatchResult, StorageMap};
+use error;
 use frame_system::{self, ensure_signed};
+use support::{debug, decl_event, decl_module, decl_storage, dispatch::DispatchResult, StorageMap};
 
 /// The CTYPE trait
 pub trait Trait: frame_system::Trait + error::Trait {
