+++ pallets/error/Cargo.toml
@@ -1,0 +1,43 @@
+[package]
+authors = [ 'KILT <info@kilt.io>' ]
+edition = '2018'
+name = 'error'
+repository = 'https://github.com/KILTprotocol/mashnet-node'
+version = '0.23.0'
+description = 'TODO'
+
+[package.metadata.docs.rs]
+targets = ['x86_64-unknown-linux-gnu']
+
+[dependencies.codec]
+default-features = false
+features = [ 'derive' ]
+package = 'parity-scale-codec'
+version = '1.3.1'
+
+[dependencies.frame-system]
+default-features = false
+version = '2.0.0-rc5'
+
+[dependencies.sp-arithmetic]
+default-features = false
+version = '2.0.0-rc5'
+
+[dependencies.sp-runtime]
+default-features = false
+version = '2.0.0-rc5'
+
+[dependencies.support]
+default-features = false
+package = 'frame-support'
+version = '2.0.0-rc5'
+
+[features]
+default = [ 'std' ]
+std = [
+	'codec/std',
+	'frame-system/std',
+	'sp-runtime/std',
+	'sp-arithmetic/std',
+	'support/std'
+]
