+++ pallets/portablegabi/Cargo.toml
@@ -1,0 +1,72 @@
+[package]
+authors = [ 'KILT <info@kilt.io>' ]
+edition = '2018'
+name = 'portablegabi'
+repository = 'https://github.com/KILTprotocol/mashnet-node'
+homepage = 'https://kiltprotocol.github.io/portablegabi-tutorial/'
+version = '0.23.0'
+description = 'RSA-B Accumulator storage for portablegabi anonymous credentials'
+
+[package.metadata.docs.rs]
+targets = ['x86_64-unknown-linux-gnu']
+
+[dependencies]
+error = { version = '0.23.0', default-features = false, path = '../error' }
+
+[dev-dependencies]
+mashnet-node-runtime = { version = '0.23.0', default-features = false, path = '../../runtime' }
+
+[dependencies.serde]
+features = ['derive']
+optional = true
+version = '1.0.111'
+
+[dependencies.codec]
+default-features = false
+features = ['derive']
+package = 'parity-scale-codec'
+version = '1.3.0'
+
+[dependencies.sp-runtime]
+default-features = false
+version = '2.0.0-rc5'
+
+[dependencies.frame-support]
+default-features = false
+version = '2.0.0-rc5'
+
+[dependencies.frame-system]
+default-features = false
+version = '2.0.0-rc5'
+
+[dependencies.pallet-balances]
+default-features = false
+version = '2.0.0-rc5'
+
+[dependencies.sp-io]
+default-features = false
+version = '2.0.0-rc5'
+
+[dependencies.sp-std]
+default-features = false
+version = '2.0.0-rc5'
+
+[dev-dependencies.sp-core]
+default-features = false
+version = '2.0.0-rc5'
+
+[dev-dependencies.sp-arithmetic]
+default-features = false
+version = '2.0.0-rc5'
+
+[features]
+default = ['std']
+std = [
+    'serde',
+    'codec/std',
+    'frame-support/std',
+    'pallet-balances/std',
+    'sp-runtime/std',
+    'sp-io/std',
+    'frame-system/std',
+]
